requirements.txt
openai==1.99.9
python-dotenv==1.0.0
aiofiles==23.2.1 
sounddevice==0.4.7
soundfile==0.12.1
numpy==2.3.3
pyperclip==1.8.2
keyboard==0.13.5
aiogram==3.13.1